Compile for Windows
- cleanup EG_API declarations, clean up Makefile, 
- add windows documentation, 
- add a c++ facade around HACL functions used in tests (using an internal api exposed in debug builds).  
- Fix string literal values in ballot nonce serialization.
- statically link  in libstdc++
-

+++ test/electionguard/test_nonces.cpp
@@ -32,7 +32,7 @@ TEST_CASE("Nonces slices produce same output as iterating for each index")
     auto n = make_unique<Nonces>(*aModQ);
     vector<unique_ptr<ElementModQ>> l1;
     l1.reserve(10);
-    for (int i(0); i < 10; ++i) {
+    for (uint64_t i(0); i < 10; ++i) {
         l1.push_back(n->get(i));
     }
     auto n2 = make_unique<Nonces>(*aModQ);
